Refactor TradeVote to Trades.TradeVote
* Refactor TradeVote to Trades.TradeVote
* See #797
