--- lib/ex338/trade_vote/store.ex
@@ -1,11 +1,0 @@
-defmodule Ex338.TradeVote.Store do
-  @moduledoc false
-
-  alias Ex338.{TradeVote, Repo}
-
-  def create_vote(attrs) do
-    %TradeVote{}
-    |> TradeVote.changeset(attrs)
-    |> Repo.insert()
-  end
-end
