+++ lib/ex338_web/admin/trade_vote.ex
@@ -3,6 +3,6 @@ defmodule Ex338Web.ExAdmin.TradeVote do
 
   use ExAdmin.Register
 
-  register_resource Ex338.TradeVote do
+  register_resource Ex338.Trades.TradeVote do
   end
 end
